+++ packages/core/src/index.ts
@@ -68,6 +68,7 @@ export * from './utils/thoughtUtils.js';
 export * from './utils/debugLogger.js';
 export * from './utils/events.js';
 export * from './utils/extensionLoader.js';
+export * from './utils/package.js';
 
 // Export services
 export * from './services/fileDiscoveryService.js';
